+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Allow customizing NFS mount options via config file
 - Updated builds to use Go 1.8
+- Allow ONTAP to split a clone from its parent upon creation
 
 ## 17.04.0